refactor: split main.rs into cmd/ submodules (Issue 2)
main.rs drops from 716 lines to 99 (dispatcher + main()). All command
implementations move to focused files under a new cmd/ submodule:

- cmd/cmd_once.rs  — run_once, resolve_once_dt, clamp_once_wall_clock_dt
- cmd/cmd_pipe.rs  — run_pipe
- cmd/cmd_loop.rs  — run_loop, LoopObserver, sleep_with_socket
- cmd/cmd_state.rs — run_status_dispatch, run_status, run_purge, run_init
- cmd/cmd_socket.rs — socket_send_and_print, run_socket_{hold,resume,reset,save,set}
- cmd/mod.rs       — module declarations, open_log, re-exports

Also fixes a pre-existing test reference to crate::app::IterationRecord
(should be pid_ctl::app::IterationRecord — the name was only resolving
because the old main.rs used pid_ctl::app::{self, ...} at the crate root).
